#19 - add watcher for rebuilding the icons json every time keywords or icons are edited
