+++ CONTRIBUTE.md
@@ -8,4 +8,22 @@ Create an issue that describes clearly the issue and the reproducible steps.
 
 #### Pull Requests
 
-If you feel the urge to contribute, submit your updates with pull request. 
+If you feel the urge to contribute, submit your updates with pull request.
+
+## Running the Doc Site
+
+First install all dependencies:
+
+```
+npm install
+```
+
+Next, run the server:
+
+```
+npm run server
+```
+
+This will optimize the icons, create a JSON file with all the icon path data, and start up a development server with live reload capabilities.
+
+This task also watches for changes and rebuilds the JSON file whenever icons are added or edited, automatically reloading your browser to display changes as they come in.
