+++ bin/cli.js
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+const build = require('./build');
+
+build(true)
+  .then(() => process.exit(0))
+  .catch(() => process.exit(1));
