+++ package.json
@@ -43,21 +43,20 @@
   "homepage": "https://github.com/Esri/calcite-ui-icons#readme",
   "scripts": {
     "spriter": "node bin/spriter.js",
-    "imageminsvgo": "node bin/imagemin-svgo.js",
     "buildsprite-16": "svgstore -o sprite-16.svg icons/*-16.svg",
     "buildsprite-24": "svgstore -o sprite-24.svg icons/*-24.svg",
     "buildsprite-32": "svgstore -o sprite-32.svg icons/*-32.svg",
     "buildsprite-16-f": "svgstore -o sprite-16-f.svg icons/*-16-f.svg",
     "buildsprite-24-f": "svgstore -o sprite-24-f.svg icons/*-24-f.svg",
     "buildsprite-32-f": "svgstore -o sprite-32-f.svg icons/*-32-f.svg",
-    "optimize": "npm run imageminsvgo",
+    "optimize": "node bin/cli.js",
     "start": "npm run buildsprite-16-f && npm run buildsprite-24-f && npm run buildsprite-32-f && npm run buildsprite-16 && npm run buildsprite-24 && npm run buildsprite-32 && npm run optimize",
-    "server": "http-server docs/"
+    "server": "node bin/server.js"
   },
   "devDependencies": {
+    "browser-sync": "^2.24.7",
     "fs-extra": "^7.0.0",
     "glob-promise": "^3.4.0",
-    "http-server": "^0.11.1",
     "imagemin": "^5.3.1",
     "imagemin-svgo": "^6.0.0",
     "promise": "~8.0.1",
